+++ CHANGELOG.md
@@ -6,6 +6,15 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 
+## [UNRELEASED]
+
+### Added
+- Introduced dedicated categories for products and products' chapters
+
+### Changed
+- Updated changelog with custom Diventi logo
+
+
 ## [1.0.0-alpha.7] → 2018-02-03
 
 ### Added